pyproject.toml
[tool.poetry]
name = "zxart-pallete"
version = "0.1.0"
description = ""
authors = ["Roman K <rmksrv@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
streamlit = "^1.38.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
